+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Changelog Format
 - [Initials of Dev] [#issue number](issue link on GitHub) {issue title}
 -->
 
+- [KF] [#6](https://github.com/KennethFranco/portfolio-v2/issues/2) layout(feature): add footer
 - [KF] [#5](https://github.com/KennethFranco/portfolio-v2/issues/2) layout(feature): add navbar
 - [KF] [#2](https://github.com/KennethFranco/portfolio-v2/issues/2) app(feature): add tailwindcss
 - [KF] [#1](https://github.com/KennethFranco/portfolio-v2/issues/1) app(feature): add initial project files